Toda validação de erro e com Postgres funcionando

+++ squadra/src/main/java/br/com/squadra/repository/Repositorio.java
@@ -1,12 +1,13 @@
 package br.com.squadra.repository;
 
+
+
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.squadra.modelo.Sistema;
 
-public interface Repositorio extends JpaRepository<Sistema, Long> {
-
+public interface Repositorio extends JpaRepository<Sistema, Integer> {
 	List<Sistema> findByDescricao(String descricao);
 }
